pyproject.toml
[project]
name = "rsa-swap"
version = "0.1.0"
description = "Add your description here"
authors = [
]
dependencies = [
    "pycryptodome>=3.20.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "rsa_swap:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.1.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rsa_swap"]
